fix: type pane and layout-cell offsets as signed, matching the C
Root cause of the crash class you kept hitting, rather than another symptom.

next-3.7 types xoff/yoff as `int` on both layout_cell (tmux.h:1276) and
window_pane (tmux.h:1518), precisely so a floating pane can be positioned
partly past an edge. ztmux stored them as u32, so `move-pane -X-10` recorded
4294967287 and every comparison derived from it either overflowed and aborted
the server or silently produced nonsense.

Both fields are now i32. Read sites cast back to u32 because the C promotes
`int + u_int` to unsigned, so the arithmetic they perform is unchanged — only
the ability to hold, print and test a negative is new. 210 read sites and 57
write sites across 18 files.

Made genuinely signed, because they compare against or derive from an offset
that can now be negative:
  - screen_write_pane_is_obscured, including the C's `xoff < 0 || yoff < 0`
    test, which was unrepresentable before
  - screen_redraw_pane_border, whose whole geometry the C computes in int
    (screen-redraw.c:460)
  - screen_redraw_draw_pane's visibility test, per-row clip and the floating
    border box
  - pane_x, pane_y, pane_left, pane_top, pane_right and pane_bottom, which the
    C prints with %d (format.c:2250, 2418, 2091, 2376)
  - layout_create_cell's sentinel, INT_MAX in the C rather than UINT_MAX

`move-pane -X-10 -Y-4` now reports @-9,-3 with L=-9 T=-3 R=18 B=2, matching
the vendored tmux exactly, and the partly-off-screen float renders
identically instead of killing the server.

Verified: 10 rendered screens byte-identical to tmux including four
off-screen placements; 18 crash scenarios with a client ATTACHED — the case
my earlier fuzzing missed by only ever calling list-panes — 0 crashes; full
cargo test, clippy and fmt clean. 1103 pins the geometry and the derived
edge formats.

parity 1130/1130.

+++ src/extensions/ratatui_ui.rs
@@ -403,10 +403,10 @@ pub(crate) unsafe fn draw_pane_number(ctx: *mut screen_redraw_ctx, wp: *mut wind
         let tty = &raw mut (*c).tty;
 
         // Prototype: draw for panes fully inside the viewport; skip partials.
-        if (*wp).xoff < (*ctx).ox
-            || (*wp).yoff < (*ctx).oy
-            || (*wp).xoff + (*wp).sx > (*ctx).ox + (*ctx).sx
-            || (*wp).yoff + (*wp).sy > (*ctx).oy + (*ctx).sy
+        if ((*wp).xoff as u32) < (*ctx).ox
+            || ((*wp).yoff as u32) < (*ctx).oy
+            || (*wp).xoff as u32 + (*wp).sx > (*ctx).ox + (*ctx).sx
+            || (*wp).yoff as u32 + (*wp).sy > (*ctx).oy + (*ctx).sy
         {
             return;
         }
@@ -453,13 +453,13 @@ pub(crate) unsafe fn draw_pane_number(ctx: *mut screen_redraw_ctx, wp: *mut wind
             }
         }
 
-        let yoff = (*wp).yoff - (*ctx).oy
+        let yoff = (*wp).yoff as u32 - (*ctx).oy
             + if (*ctx).statustop != 0 {
                 (*ctx).statuslines
             } else {
                 0
             };
-        let xoff = (*wp).xoff - (*ctx).ox;
+        let xoff = (*wp).xoff as u32 - (*ctx).ox;
         let px = xoff + ((*wp).sx.saturating_sub(boxw as u32)) / 2;
         let py = yoff + ((*wp).sy.saturating_sub(boxh as u32)) / 2;
         blit_tty(tty, &buf, px, py);
@@ -1546,10 +1546,10 @@ pub(crate) unsafe fn draw_pane_frame(ctx: *mut screen_redraw_ctx, wp: *mut windo
         if (*wp).sy <= 2 {
             let width = (*wp).sx as usize;
             if width < 4
-                || (*wp).yoff < (*ctx).oy
-                || (*wp).yoff >= (*ctx).oy + (*ctx).sy
-                || (*wp).xoff < (*ctx).ox
-                || (*wp).xoff + (*wp).sx > (*ctx).ox + (*ctx).sx
+                || ((*wp).yoff as u32) < (*ctx).oy
+                || ((*wp).yoff as u32) >= (*ctx).oy + (*ctx).sy
+                || ((*wp).xoff as u32) < (*ctx).ox
+                || (*wp).xoff as u32 + (*wp).sx > (*ctx).ox + (*ctx).sx
             {
                 return;
             }
@@ -1587,13 +1587,13 @@ pub(crate) unsafe fn draw_pane_frame(ctx: *mut screen_redraw_ctx, wp: *mut windo
             for (x, ch) in row.iter().enumerate() {
                 buf[(x as u16, 0)].set_char(*ch).set_style(st);
             }
-            let yoff = (*wp).yoff - (*ctx).oy
+            let yoff = (*wp).yoff as u32 - (*ctx).oy
                 + if (*ctx).statustop != 0 {
                     (*ctx).statuslines
                 } else {
                     0
                 };
-            let xoff = (*wp).xoff - (*ctx).ox;
+            let xoff = (*wp).xoff as u32 - (*ctx).ox;
             let tty = &raw mut (*(*ctx).c).tty;
             blit_tty(tty, &buf, xoff, yoff);
             return;
@@ -1601,8 +1601,8 @@ pub(crate) unsafe fn draw_pane_frame(ctx: *mut screen_redraw_ctx, wp: *mut windo
 
         // The frame occupies the reserved ring: the full layout cell, one inset
         // cell out from the content on each side.
-        let cell_x = (*wp).xoff.saturating_sub(inset);
-        let cell_y = (*wp).yoff.saturating_sub(inset);
+        let cell_x = ((*wp).xoff as u32).saturating_sub(inset);
+        let cell_y = ((*wp).yoff as u32).saturating_sub(inset);
         let cell_sx = ((*wp).sx + 2 * inset) as u16;
         let cell_sy = ((*wp).sy + 2 * inset) as u16;
         if cell_sx < 4 || cell_sy < 3 {
